Remove imprecise and unnecessary HOME ownership check.
When looking for the pulse home directory, we currently check to see
if HOME is owned by us. This check is imprecise, because what we
actually care about is being able to create a subdirectory of ~/.pulse
or $XDG_CONFIG_HOME/pulse, and unnecessary, because if we don't have
the appropriate permissions we will get an appropriate failure when we
try to create the subdirectory immediately after this function
call (and the normal TOCTOU concerns apply).

This check was added in a9c80b43831411fac4071ea1cb3e23bd42e6bd9d, and
nothing in the commit message indicates why it would be needed.