qpaeq: fix: resizing on HiDPI
On high-dpi setups resizing qpaeq caused infinite recalculations because `event.oldSize().width()` returned irrelevant values that resulted in "thousands of sliders desired". Using different approach that may be not as accurate, but works on HiDPI setups without issues.

Long-running continuous resizing caused recalculations to be done for each event that was not really necessary.