virtual sinks/sources: Propagate asyncmsgq change after sink-input/source-output move
Currently, when the master of a virtual sink/source is moved, the change of the
asyncmsgq is not propagated to other attached virtual sinks/sources. This leads
to a crash when the original master sink/source is no longer available.

This patch fixes the issue by modifying the moving callback to propagate the
change to attached virtual sinks/sources.