launch: Add pulseaudio-x11 systemd user session service
This is a oneshot service that may be used to tell Pulseaudio to load
or unload modules that rely on X11 display availability. Starting or
stopping it will result in the cork/x11publish/xdmp modules being
added/removed.

The start-pulseaudio-x11 script has been made to indirectly rely on
this script, after initial checks.

With systemd activation, the session will be able to load these
modules either right away (eg. on actual X11 sessions), or in a
delayed manner (eg. after Xwayland is started).