context: Fix some annotations from constification
pa_context_set_error() mutates the error inside the context, so the
const annotation is incorrect.

Some of the other API that was marked as having a const pa_context* as
input was actually modifying the the error on the context while doing a
validity check. For those cases also, we must drop the const qualifier.