source-output: Fix rewinding
If the output implements a process_rewind() callback, the resampler delay is
not taken into account. This leads to glitches during volume changes when
source and source output rates differ.

This patch fixes the problem. Additionally, the helper functions
pa_convert_to_source_length() and pa_convert_to_source_output_length()
are introduced, similar to the functions for the sink input side. The
pa_convert_to_source_length() function is unused and only provided for
completeness.