bluetooth: Set all new streams to suspended state and remove argument "optional" from transport acquire function
Now when broken local AudioGateway role was removed there is no need to
have special (and moreover non-working) hacks for it in transport code.
This would allow to create streams for all transports immediately in
suspended state, so applications can bind to them.

A2DP source profiles can be also initialized via "Acquire" method, so there
not need to "TryAcquire" method for them. Special handling is needed only
for SCO master sockets which are not handled by pulseaudio anymore.
Moreover is is possible to do it in smarter way, like in hsphfpd daemon.