switch-on-port-available: check if need to change the default sink or source
After a port is selected, if it is in the current default_sink or
default_source, the port could be set to the current output/input
device.

But if the port is not in the current default_sink or
default_source, and it is in the same profile with the current
default_sink or default_source, the port will not be set to
the current output/input device, this doesn't meet the
users' expectation. For example:
On the machine with the sof driver (dmic connects to the PCH instead
of the codec), the external mic belongs to one source while the dmic
belongs to another source, before plugging in the external mic, the
dmic is the default_source->active_port, and both external mic and
dmic belong to the same profile HiFi. After plugging in the external
mic, the external mic is set to the active_port of its source, but the
source of external mic is not set to be the default_source, so the
default_source->active_port is still the dmic.

After applying this patch, the default_source or default_sink will be
changed as well if needed.

BTW, if the port already belongs to the default_sink or
default source, it is harmless to call set_default_sink/source again.

Signed-off-by: Hui Wang <hui.wang@canonical.com>